requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-telegram-bot==21.3
pydantic==2.7.1
